Porting to python3(Heavy WIP)

+++ lbrynet/dht/distance.py
@@ -12,10 +12,10 @@ class Distance(object):
         if len(key) != constants.key_bits / 8:
             raise ValueError("invalid key length: %i" % len(key))
         self.key = key
-        self.val_key_one = long(key.encode('hex'), 16)
+        self.val_key_one = int(key.encode('hex'), 16)
 
     def __call__(self, key_two):
-        val_key_two = long(key_two.encode('hex'), 16)
+        val_key_two = int(key_two.encode('hex'), 16)
         return self.val_key_one ^ val_key_two
 
     def is_closer(self, a, b):
